# REZ-api-gateway/src/openapi.yaml
openapi: 3.0.3
info:
  title: REZ Intelligence API
  description: |
    AI-powered platform for local commerce with 3 client types:
    - REZ_ECOSYSTEM: Full intelligence sharing
    - NON_REZ: Strict tenant isolation
    - RABTUL_SAAS: Plugin-based SaaS
  version: 1.0.0
  contact:
    name: REZ Engineering
    email: engineering@rez.money
    url: https://rez.money
servers:
  - url: http://localhost:4300
    description: Local development
  - url: https://rez-api-gateway.rez.money
    description: Production
...
paths:
...
  /api/profile/{userId}:
    put:
      tags: [Profile]
      summary: Update user profile
      security:
        - ApiKeyAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'

  # Workflows (via Flow Runtime)
...
  /api/workflows/{workflowId}:
    get:
      tags: [Workflows]
      summary: Get workflow
      security:
        - ApiKeyAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for tenant authentication.
        Format: {prefix}_{tenant}_{id}
        - REZ_ECOSYSTEM: rez_*
        - NON_REZ: ext_*
        - RABTUL_SAAS: saas_*

    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Internal service-to-service authentication
  schemas:
...
    UserProfile:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        avatar:
          type: string
          format: uri
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        segments:
          type: array
          items:
            type: string
        tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        lifetimeValue:
          type: number
        memberSince:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        cuisine:
          type: array
          items:
            type: string
        priceRange:
          type: array
          items:
            type: string
            enum: [budget, medium, premium]
        notificationPreferences:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            sms:
              type: boolean
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        avatar:
          type: string
        preferences:
          $ref: '#/components/schemas/UserPreferences'

    # Workflows
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, paused, archived]
        entryNodeId:
          type: string
        nodeCount:
          type: integer
        executionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...